Rename `single-user` to `standalone`
