+++ docker/.env
@@ -282,8 +282,8 @@ MF_TWINS_SERVER_KEY=""
 MF_TWINS_DB=mainflux-twins
 MF_TWINS_DB_HOST=twins-db
 MF_TWINS_DB_PORT=27018
-MF_TWINS_SINGLE_USER_EMAIL=""
-MF_TWINS_SINGLE_USER_TOKEN=""
+MF_TWINS_STANDALONE_EMAIL=""
+MF_TWINS_STANDALONE_TOKEN=""
 MF_TWINS_CLIENT_TLS=""
 MF_TWINS_CA_CERTS=""
 MF_TWINS_CHANNEL_ID=
