+++ things/standalone/doc.go
@@ -1,0 +1,9 @@
+// Copyright (c) Mainflux
+// SPDX-License-Identifier: Apache-2.0
+
+// Package standalone contains implementation for auth service in
+// single-user scenario. Running with a single user provides
+// Things as a standalone service with one admin user who
+// manages all the Things and Channels and does not
+// require connection to Auth service.
+package standalone
